backend/recipes/requirements.txt
cycler==0.11.0
Django==2.2.19
django-filter==21.1
djangorestframework==3.13.1
fonttools==4.33.3
gunicorn==20.1.0
kiwisolver==1.4.2
numpy==1.21.6
packaging==21.3
Pillow==9.1.0
psycopg2-binary==2.9.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
